# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Forgotten Realms City Generator API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface CityInput {
  /** @maxLength 25 */
  name: string;
  size: string;
  region: string;
  culture: string;
  alignment: string;
}

export interface NpcAtributos {
  for: string;
  des: string;
  con: string;
  int: string;
  sab: string;
  car: string;
}

export interface NpcEstatisticas {
  nivel?: string;
  bonusProficiencia: string;
  classeArmadura: number;
  pontosDeSaude: string;
  atributos: NpcAtributos;
  pericias: string[];
}

export interface Npc {
  nome: string;
  raca: string;
  classe: string;
  descricao: string;
  papel: string;
  segredo: string;
  estatisticas?: NpcEstatisticas;
}

export interface City {
  nome: string;
  alcunha: string;
  resumo: string;
  populacao: string;
  geografia: string;
  economia: string;
  governo: string;
  marcasFamosos: string;
  culturaLocal: string;
  problemas: string;
  npcs: Npc[];
}

export interface DistrictInput {
  tipo: string;
  /**
     * @minimum 1
     * @maximum 10
     */
  estabelecimentos: number;
  /**
     * @minimum 0
     * @maximum 5
     */
  bases: number;
  cityContext: string;
}

export interface District {
  nome: string;
  tipo: string;
  atmosfera: string;
  descricao: string;
  populacao: string;
  funcao: string;
  estabelecimentos: string[];
  bases: string[];
  problemasLocais: string;
  npcs: Npc[];
}

export interface NpcInput {
  cityContext: string;
  /** @nullable */
  districtContext?: string | null;
}

export interface RumorsInput {
  cityContext: string;
}

export interface Rumors {
  rumores: string[];
}

export interface AdventureHookInput {
  cityContext: string;
  /** @nullable */
  districtsContext?: string | null;
}

export interface AdventureHook {
  titulo: string;
  descricao: string;
  complicacao: string;
  recompensa: string;
}