# lib/api-zod/src/generated/types/npc.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Forgotten Realms City Generator API
 * OpenAPI spec version: 0.1.0
 */
import type { NpcEstatisticas } from './npcEstatisticas';

export interface Npc {
  nome: string;
  raca: string;
  classe: string;
  descricao: string;
  papel: string;
  segredo: string;
  estatisticas?: NpcEstatisticas;
}