# lib/api-zod/src/generated/types/npcAtributos.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Forgotten Realms City Generator API
 * OpenAPI spec version: 0.1.0
 */

export interface NpcAtributos {
  for: string;
  des: string;
  con: string;
  int: string;
  sab: string;
  car: string;
}