# lib/api-zod/src/generated/types/npcEstatisticas.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Forgotten Realms City Generator API
 * OpenAPI spec version: 0.1.0
 */
import type { NpcAtributos } from './npcAtributos';

export interface NpcEstatisticas {
  nivel?: string;
  bonusProficiencia: string;
  classeArmadura: number;
  pontosDeSaude: string;
  atributos: NpcAtributos;
  pericias: string[];
}